Return :nil on read of non-existent document
Jepsen assumes all reads which return nil value are failed reads, even
if the status is ok. This prevents us from distinguishing between failed
reads and reads where all mutations for a key have been lost. Instead,
return the keyword :nil when the key doesn't exist, and set this to
be the initial value.