Rework client code to use SDK 3.0.0 alpha
The new java sdk 3.0 allows us to perform syncwrites. Add the couchbase
preview repository to the project configuration and use the 3.0.0-alpha.1
build of the new sdk. Rework and simplify the client code for the new sdk.

Note that the DCP client still uses the "old" core-io library, which
conflicts with the "new" core-io library used by the new java client.
The DCP client therefore needs to be manually recompiled with patches
that shade the core-io dependency, then installed to the control node's
local maven cache.

Change-Id: I0de8b11265c17fc42e696e8e4fa26ef5a689a9e3
Reviewed-on: http://review.couchbase.org/106080
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Tested-by: Korrigan Clark <korrigan.clark@couchbase.com>