Obey timeout parameter
Currently the --durability-timeout parameter is accepted during argument
parsing but does not take effect. The Java SDK does not differentiate
between the client side timeout applied to kv operations and the server
side timeout specified in the durability flexframe of sync-write request
headers. Apply the "durability-timeout" as the timeout for all kv
operations and rename it to --kv-timeout to reflect this. Modify the
suite files to use the new parameter.

Change-Id: Iddd02475abbed9202d674e3e27f3d2805df252a0
Reviewed-on: http://review.couchbase.org/106944
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Tested-by: Korrigan Clark <korrigan.clark@couchbase.com>