Correctly reflect that the dcp-client jar is based off 0.23.0
When I first built the dcp-client jar and merged it into this repo,
I listed it as built off 0.21.0. In fact it was really built off
0.23.0 code. I've build it again off exactly 0.23.0 and am readding
it to this repository under the correct name.

The pom file to regenerate the dcp-client jar can be found here:
http://review.couchbase.org/#/c/110267/.

Additionally, I'm reworking the dependencies into a more natural
order with more basic dependencies first and more complex
dependencies later.

Change-Id: Ib32a91a6a5e7982ea5c84c98300749902c609113
Reviewed-on: http://review.couchbase.org/110328
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>