MB-34656:Update run.sh write Jepsen output to a file
Update run.sh to pipe output of the tests to a log file. Otherwise
the console output becomes to large of the jenkins console to handle.
Add a flag to enable the renaming of tests results in the store
directory from ./store/Couchbase to ./store/Couchbase-$build-job-name.
As this helps to distinguish between jobs for the results on the
KV-Engine CV machines when view results on the Jepsen log ui.

Change-Id: I1a8fe136653d2cb85b652e848a9c8f8a6fb546cf
Reviewed-on: http://review.couchbase.org/111093
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Sven Signer <sven.signer@couchbase.com>