MB-34656: Move to Jepsen 0.1.14
Move Jepsen to 0.1.14, to do this we need to update our cli for Jepsen
and update our checkers not to use models as Jepsen's checkers no-longer
support models as part of the checker interface. Also add
org.clojure/tools.cli as an include to add support for :default-fn for
our jepsen cli.

Change-Id: I779d84254e3f76e743ccd818309944e39c94b42a
Reviewed-on: http://review.couchbase.org/111440
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Sven Signer <sven.signer@couchbase.com>