MB-35129 Move devmapper setup to util/setup-node
Restarting Couchbase Server to move the data path onto a devmapper
device during nemesis setup appears to be causing issues for the
clients. Originally we performed the devmapper setup during the
nemesis setup so we could automatically detect if it was necessary.
Now that the --manipulate-disks parameter is needed anyway, it is
cleaner to perform devmapper setup during the actual node setup.
This avoids the need to restart Couchbase Server, ensuring the nodes
are up and ready when the clients are connecting.

Change-Id: I184372138dfdd2dcb1ba87834de061fe0bd80c2a
Reviewed-on: http://review.couchbase.org/112103
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>