MB-35043 Move OS setup to vm-setup script
We currently still use Jepsen's OS setup to install ntpdate on the our
VMs. This was introduced to install this dependency on simple vagrants,
but as we add more options it increases overhead for this logic. Move
this to the vm-setup script in preparation for changes to our Jepsen
node setup when implementing cluster-run style nodes in Jepsen.

Change-Id: I757452320d3f440e3a49666ff15abb3e8d4b2f36
Reviewed-on: http://review.couchbase.org/113271
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>