MB-35291 Move to GET based set read
Change the DCP workloads such that the final read is performed by
executing standard MCBP GET requests for each attempted key. Previously
we could not be sure if a key was not streamed due to it being lost, or
simply due to the server not being ready to stream the key. By using
MCBP requests rather than DCP, we can catch TMPFAIL and similar errors.
The previous behaviour is still available by using the --dcp-set-read
option.

Change-Id: Ibd83f5540a541f467258737c303e1857053ff1ef
Reviewed-on: http://review.couchbase.org/113568
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>