MB-35292 Fix observe based durability
Fix the observe based durability handling such that the replicate-to
and persist-to parameters can be passed and take effect. Also remove
the JDK-4283544 workaround as internal changes in newer alpha builds
of the Couchbase Java SDK mean that this issue is no longer triggred
when invoking the methods in the usual way.

Change-Id: I3ad938aedef529b21f2beeeb305d88f60579c41e
Reviewed-on: http://review.couchbase.org/113374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>