MB-35291 Move to GET based set read (re-commit)
This is a re-commit of previously reverted commit 2a2dfc4
(http://review.couchbase.org/#/c/113568/) following a fix to an
invalid test condition when determining whether to construct a
dcp-client in the set-kill workload.

Original commit message:

Change the DCP workloads such that the final read is performed by
executing standard MCBP GET requests for each attempted key. Previously
we could not be sure if a key was not streamed due to it being lost, or
simply due to the server not being ready to stream the key. By using
MCBP requests rather than DCP, we can catch TMPFAIL and similar errors.
The previous behaviour is still available by using the --dcp-set-read
option.

Change-Id: I1c7c9bb42179392bd81f19017d9f4ab393628dba
Reviewed-on: http://review.couchbase.org/113645
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>