MB-35752 Send rebalance rest-call to current control host
In upcoming commits, we will need to be able to control which node
serves as the rest-target during a rebalance. Modify the behaviour
of rebalance such that the current Jepsen control host is used
instead of automatically picking one. Note that this commit breaks
the rebalance nemeses until the remainder of the MB-35752 patches
have been applied.

Change-Id: I7ff977e08c8772d42962c9f99355ed35956256c2
Reviewed-on: http://review.couchbase.org/114063
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>