MB-35759 Fix disk-failure workload
The disk-failure was previously broken, both in implementation and
by design. Modify the workload such that it no longer attempts to
perform any recover following the disk failure. Since the simulated
failure is performed at the block device level, we break the file
system on which the data is stored. We can't recover from this,
don't try to rebalance nodes back into the cluster. Note that this
also means we can't perform multiple cycles. Also fix implementation
such that the workload works correctly with the new targeting logic.

Also add new disk-failure workload tests to the nightly and
post-commit Jenkins jobs.

Co-Authored-By: Sven Signer <sven.signer@couchbase.com>
Co-Authored-By: Richard de Mellow <richard.demellow@couchbase.com>

Change-Id: Ie9c12c26763ffeff05be49fd34942fc11818eb47
Reviewed-on: http://review.couchbase.org/117294
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>