MB-37398: Add counter workload that handles inc and dec
Add a counter workload that can perform increments and decrements. To
support this also introduce a sanity-counter checker that dose not
check the counters value though time but instead checks its value at the
end of the test against what we should have seen based on successful
ops. Also ensure that we can change the level of durability of our ops
from the cli.

Change-Id: I6fa7328f5737f318cc157b3f4050c6c523bf7bab
Reviewed-on: http://review.couchbase.org/120493
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>