MB-37337: Add lein uberjar support
Add lein uberjar support to allow us to create a jar file that will run
our tests directly with java. To do this remove the java-dcp-client lib
dir as a resource and instead use maven to get the library. As part of
this also update the version of java-dcp-client, which requires some
small updates to the api that we use.

Change-Id: Iae00e628059a7987c4ee8ac675014e2cf05c4770
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/134146
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>