Fix CLI options --use-subdoc and --doc-padding-size
Fix CLI options use-subdoc and doc-padding-size, as currently we do not
update *use-subdoc* correctly. To do this ensure *use-subdoc* is
modified using parse option function. doc-padding-size is floored due
to (create-int-json-obj) not being able to access "extra-cli-options"
as its not in the static name space. Thus, make it global variable so it
can be accessed from the global static static name space.

Change-Id: I7b9788a22a95839be0c05c80317434bd637b778a
Reviewed-on: http://review.couchbase.org/122785
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>