MB-37395: Modify the logic of checking rollbacks in upset-checker
Currently, upsert-checker checks if there was an add-rollback(and
upsert-rollback) for both indeterminate-add/upsert and failed-add/upsert
scenarios. This can create a problem because an indeterminate op's
result is not known. Upsert checker fails to distinguish a failed-op
(type:fail) and an indeterminate-op (type:info) currently. So this
change modifies upset-checker to distinguish failed-op and
indeterminate-op first and checks for rollbacks only on failed-ops.
Accordingly, the test for this checker is also modified.

Change-Id: Ifada3c5e461497e1d9ea3e26b0e6cfedc21ee8cd
Reviewed-on: http://review.couchbase.org/123579
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>