MB-38165: Gzip pcap files after each test run
Currently after a test run we store pcaps uncompressed, these can be GB
which rapidly uses up disk space. To try and reduce this we should
compress all *.pcap files after each test run as this can give us ~10x
reduction in disk storage for these files. So, in the get-remote-logs
function, we run a gzip command to compress all the pcap files before
moving it to /tmp/jepsen-logs folder.

Change-Id: I0d94c5fa77f2e51e49d3ebb8a93c5cb78f040bcb
Reviewed-on: http://review.couchbase.org/123829
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>