Move DGM test for post-commit from failover to kill workload
Move DGM test for post-commit conf from the failover workload to the
kill workload. As currently the DGM test makes the failover rebalance
last a long time. This causes our complexity for analysis to increase
past a point were we can not longer perform analysis of the results.
Which means the test isn't useful thus, if we move it to a faster
running test then we get a DGM test that we can analyse.

Change-Id: I883de412574547a46c609ce86700567f3cde20cb
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/124521
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>