MB-37367 Update Jepsen dependency to 0.1.19
Jepsen 0.1.19 is the last release in Jepsen's 0.1.x series. Bump our
Jepsen dependency to this version in preparation for future commits to
incrementally address compatability issues with breaking changes
introduced by Jepsen 0.2.0.

An earlier attempt at updating Jepsen from 0.1.15 to >=0.1.16 caused
strange dependency issues, this appears to fixed for now. See the MB
for further details.

Previously we were using Clojure 1.10.1, which is ahead of the internal
Clojure dependency defined by Jepsen <=0.1.19. This causes a crash with
Jepsen 0.1.19, which seems to be related to Jepsen's new generator code
and is fixed in Jepsen 0.2.0 (see Jepsen commit 6d21425). Looking at
Clojure's changelog for 1.10.1 it doesn't appear that any of the
mentioned changes affect us so it should be ok to drop down to Clojure
1.10.0 until we can bump Jepsen to 0.2.0.

Change-Id: I1ead5c0717f1dfaa398e7cd1491638e5fe601d4f
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/133363
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>