MB-37399 Add Sequential Consistency Checker
Add a new sequential consistency checker. Initial testing suggests that
the checker is working correctly, but I'm not completely sure all edge
cases are covered.

We test for sequential consistency by attempting to build a single-order
of all operations that obeys PRAM characteristics. In order to simplify
this checker we limit the set of valid histories to those with
exclusively :read and :write operations (no CAS), and only allow a
single :write for a given value. By assuming RVal characteristics we can
then always trace an observed read value to the corresponding write op.

Performance is somewhat mixed. For plausible real histories this
checker should perform roughly n*t operations for t doc-threads and n
total operations in the history. So far we seem to be consuming large
amounts of RAM proportional to history length, causing somewhat of a
cliff edge for performance once we start hitting swap. This can likely
be greatly optimized.

Change-Id: Ia5232d329328f2c6c0e6fdb30dd7d9316688ca10
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/126161
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>