MB-37436 (15/n) Prepare for set workloads
Add util functions and update client code to allow for namespaced
set workloads. When using pure generators clients no longer have
access to the test history, so we need to manually track the
largest attempted add operation and pass this to the set client
such that it knows which keys to probe.

Change-Id: I2a15ad34b84069be14f224d68d3eab6618b2bf57
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/137016
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>