MB-37399 Add CLI options to control checker
Add a new CLI option to control which checker(s) are used for register
workloads. The --use-checker option activates a checker, and can be
specified multiple times to activate multiple checkers. If the option
is not specified we default to checking linearizability.

Note that enabling the sequential consistency checker causes workloads
to write monotonically increasing integer values rather than random
values to satisfy the checker's requirement for unique write values.

Change-Id: I45eaf3220b4f08187dd9c2e9449e0335e521fbfe
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/130638
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>