MB-41772 Fix set-delete client generator
The set-delete workload client generators are currently broken,
yielding only 45 operations by default. Modify the generator such
that the intended default of 10000 keys is applied. Further modify
the behaviour of the client generator such that it can cycle
indefinitely on new batches of keys until the nemesis has completed,
bringing it in line with the behaviour of other workloads.

Change-Id: I9695ae1d936f4acc850550dbf38a9d1213fb543e
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/137255
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>