Make run.sh not treat CouchKVStore::initialize errors as failures
Currently we will treat error logs in memcached as reason to fail the
test. But for the kill-n-disk-failure workload we expect to see
CouchKVStore::initialize errors as after we've kill memcached and the
disk, couchstore should not be able to read its data files. So we want to
remove these log lines from any other errors messages before we decided
to fail the test.

Change-Id: I9e5eb1027d783f0fc5fe74a03f4b690fcf19ab6c
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/146417
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>