MB-45214:Don't fail the build if Prometheus failed to start
Don't fail the build if Prometheus-cpp logs a CRITICAL/ERROR due to it
not being able to start as it can't connect to port 11280. We expect
this situation to occur when we've killed the babysitter. As it will
shutdown and re-start memcached and Prometheus. However, its likely that
Prometheus will try and restart before memcached has fully shutdown.
Thus, meaning the port will still be in use.

Change-Id: Ic258d9b47a4687b47c8d6345b1a3dc8c012b7f9c
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/149523
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>