MB-46033: Remove transactions related code
This patch removes the client code, command line options and libraries
related to transactions.

The transactions related code is in an intermediate stage where it would
require a generator and a model for it to function. If completed, the
code would perhaps be leaning towards testing for strict
serializability.

As the aim is to test for transactional models below or equal to the
level of serialisability, the existing transactions code is being
removed with plans to reintroduce it in the future with the amended
semantics.

Change-Id: I8beaaf488b71b3aaf2fe828744ec775167a9f20c
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/152791
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>