MB-45570: Switch to password based authentication
Although the private key used by the vagrants is publically available,
we can switch to password based authentication to avoid having one in
the repository.

Changed the credentials of the vagrant machines to be consistent with
Jepsen's default credentials of 'root:root'.

Tests will now run using the root user as opposed to the vagrant user.

Jepsen will use these credentials by default when no --ssh-private-key
argument is provided.

Updated the run.sh script and the README to omit the usage of the
--ssh-private-key and --username argument.

Change-Id: If948a7406490a973ee8d864ecb64ea6400b67ca3
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/152917
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>