MB-46182: Reduce replicas in swap-rebalance test
Changed the number of replicas from 2 to 1 in the swap-rebalance test in
the nightly commit validation test-suite.

The CV machines being underspecced resulted in perpetual auto-compaction
slowing down the progress of the rebalance consequently crashing the
rebalance nemesis process.

Reducing the number of replicas increases the speed of auto-compaction
which will in turn allow the rebalance to make progress.

Change-Id: I55784ae2436c9e453ba29ff9aa96d8948bcac694
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/153146
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>