MB-46220: Retry mechanism for adding nodes
Previously, a fixed sleep period of 5 seconds was insufficiently long
for the add node to succeed following a rebalance consequently resulting
in various rebalance nemesis crashing.

As add-node is an indempotent operation on failure, we can retry with an
exponential back-off period as a less naive approach in comparison to
simply increasing the sleep time.

Change-Id: I73a964578146ff9e31d1875521d61d6d65e0a6e7
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/153350
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>