MB-37438: Refactor clients.clj to unify catch list
The exceptions thrown by read and write operations can be broken down
into two categories. Operations that are idempotent (read) and
operations that are not (e.g. write and cas).

This patch is a refactor of these two cases by introducing macros taking
an op and a body. If a known exception is thrown the op is populated
with the :type and :error fields set.

The addition of the macros offers a single place to make changes to
various exceptions consequently minimising duplication.

On CouchbaseExceptions:

The final catch in the macros is the CouchbaseException which is the
parent class of all exceptions raised by the SDK.

When raised:
i) For the idempotent case, it is a fail as reads are idempotent.
ii) For the non-idempotent case, it is as an indeterminate result.

Removed the catch for RuntimeException in do-counter-add.  Previously
this was added to deal with the changing SDK as a catch all approach.
Catching CouchbaseException should be sufficient now as the catch all
approach.

Change-Id: I29f7f62f18a13284f020be9a858cc1ce3c4a7a05
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/153743
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>