MB-46583: Refactor rest-call method
Refactored the rest-call method to allow more flexibility in terms of
what type of http request can be made and options that can be supplied
to underlying http library.

In particular, it now is possible to supply a http method (e.g. :post,
:get and :put) and supply various additional optional arguments through
the usage of a dictionary as the last argument.

For instance, you can specify the :target of the http request, the http
:params and additional :options which configure the type of data (e.g.
json as opposed to the default).

Change-Id: I720bb6548adb3150b7d59400ac7582728a4552fa
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/154510
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>