MB-46859: Update Jepsen from 0.2.3 to 0.2.4
This patch bumps the Jepsen from 0.2.3 to 0.2.4 and switches to ssh key
based authentication.

Jepsen 0.2.4 shells out to scp as opposed to using a ssh library which
significantly speeds up uploads and downloads.

However, Jepsen 0.2.4 is no longer backwards compatible with:
1. Username and password based authentication.
2. Acknowledging the global StrictHostKeyChecking settings.

Consequently, this patch introduces as part of the provisioning phase, a
script 'ssh_keys.sh' which facilitates:
1) Creation of ssh key pairs.
2) Insertion of fingerprints into the known_hosts file for nodes.
3) Publishing of the ssh public keys to nodes given their credentials.

Change-Id: Ib32140d71aefa8ff0075558a31131b8f1e70564f
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/155464
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>