MB-47937: Add a retry around /controller/addNode
To address MB-45289, in which the '/controller/addNode' end-point may
return status code of 500, added a retry with exponential backoff inside
the add-node function given this error is retryable.

As a pre-requisite to this, refactored the http request portion of
add-nodes into a new function add-node to improve readability.

Finally, removed each occurence of retry-with-exp-backoff from each
add-nodes call in the nemesis namespace.

As a side effect this reduces duplication as each occurence of add-nodes
no longer needs to be surrounded with a retry-with-exp-backoff.

Change-Id: I5b866bfaceb86d843d0588734c7955cbe58da401
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/159380
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>