MB-47950: Move retry inside failover function
Move the retry-with-exp-backoff from outside the failover function call
to inside the failover function.

This is because semantically we are retrying on the http status code as
opposed to failover function itself raising an exception.

As a side effect, this also reduces repetition as each occurence of the
function no longer has to be surrounded with a retry-with-exp-backoff.

Change-Id: I7532425eae86954e67e24a635b3bfaa4d43708c3
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/159381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>