MB-48095: Add the ability to purge logs of a test that passes
Add the ability for run.sh to purge any logs collected apart from
results.edn if the test passes to save disk space. We do not remove
results.edn as it allows the Jepsen web server to display that the
test passed.

This patch also modifies kv-cv-job.sh to use this functionality for CV
jobs.

Change-Id: I621ec57c3cfe4d322d6fea70c4d3c6131fff478b
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/159974
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Asad Zaidi <asad.zaidi@couchbase.com>