MB-40907: Add Magma as a possible backend storage
The aim of this patch is to enable Magma as a possible choice for the
bucket's storage backend.

Added a --storage-backend parameter which accepts 'magma' and (by
default is set to) 'couchstore'.

Updated create-bucket to supply the 'storageBackend' parameter when the
--storage-backend is set to 'magma' compatability. To maintain backwards
compatability when the --storage-backend is set to 'couchstore', this
parameter is not sent as part of the http request.

To improve readability, refactored out the code which selects the bucket
parameter from test data into a separate function.

Change-Id: I1d2c67b5441416ec388bb3e3e9262f6f97b4fbfe
Reviewed-on: http://review.couchbase.org/c/jepsen.couchbase/+/161965
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>